requirements.txt
asteroid-filterbanks==0.4
einops==0.6.0
numpy==1.24.0
pandas==1.3.0
rich==12.0.0
torch_audiomentations==0.11.0
torchmetrics==0.11.0
ml_collections==0.1.1
wandb==0.15.12
s3prl==0.4.11
speechcolab==0.0.6a0
jiwer==3.0.3
-e '../lhotse[dev]'
# pip install setuptools==59.5.0